+++ duck_orm/model.py
@@ -1,22 +1,31 @@
-from typing import Dict, List, Tuple, Type, TypeVar
+from typing import Any, Dict, List, Tuple, Type, TypeVar
 from databases import Database
 import inspect
 
 from duck_orm.utils.functions import get_dialect
-from duck_orm.exceptions import UpdateException
+from duck_orm.exceptions import IdInvalidException, UpdateException
 from duck_orm.sql import fields as fields_type
 from duck_orm.sql.condition import Condition
 
 T = TypeVar('T', bound='Model')
 
 
-class Model:
+class ModelMeta(type):
+    def __new__(cls, name, bases, attrs):
+        model_class = super().__new__(cls, name, bases, attrs)
+
+        if ("model_manager" in attrs):
+            attrs['model_manager'].add_model(name, model_class)
+
+        return model_class
+
+
+class Model(metaclass=ModelMeta):
     __tablename__: str = ''
     __db__: Database
 
     def __init__(self, **kwargs):
         self._instance = {}
-        self.relationships()
 
         for key, value in kwargs.items():
             self._instance[key] = value
@@ -36,22 +45,37 @@ class Model:
     def __getitem__(self, key):
         return getattr(self, key)
 
-    def relationships(self):
+    @classmethod
+    def relationships(cls):
         pass
 
     @classmethod
     async def associations(cls):
         cls.relationships()
         sqls: list[str] = []
-        for _, field in inspect.getmembers(cls):
+        dialect = cls.__db__.url.dialect
+
+        for name, field in inspect.getmembers(cls):
             if isinstance(field, fields_type.Column):
-                from duck_orm.sql.relationship import OneToMany
-                if isinstance(field, OneToMany):
-                    dialect = cls.__db__.url.dialect
-                    field_name, field_id = cls.get_id()
-                    sql = field.sql(dialect, cls.get_name(), field_name,
-                                    field_id.type_sql(dialect))
-                    sqls = sqls + sql
+                from duck_orm.sql.relationship import ForeignKey, OneToOne
+                sql = ''
+                if isinstance(field, ForeignKey):
+                    field_id = field.model.get_id()[1]
+                    sql_field_fk = field_id.type_sql(dialect)
+                    sql = field.sql(
+                        dialect=dialect, name=name, table_name=cls.get_name(),
+                        type_sql=sql_field_fk)
+                elif isinstance(field, OneToOne):
+                    if dialect == 'sqlite':
+                        await cls.drop_table()
+                        await cls.create()
+                    elif dialect == 'postgresql':
+                        field_id = field.model.get_id()[1]
+                        type_sql = field_id.column_sql(dialect)
+                        sql = field.sql(
+                            dialect=dialect, field_name=name,
+                            type_sql=type_sql, table_name=cls.get_name())
+                sqls.append(sql)
 
         for sql in sqls:
             await cls.__db__.execute(sql)
@@ -67,30 +91,17 @@ class Model:
     def __get_create_sql(cls):
         fields: List[Tuple[str, str]] = []
         sql_unique_fields = []
+        dialect = cls.__db__.url.dialect
 
         for name, field in inspect.getmembers(cls):
             if isinstance(field, fields_type.Column):
-                dialect = cls.__db__.url.dialect
-                from duck_orm.sql.relationship import (
-                    OneToMany, ManyToOne, OneToOne, ForeignKey)
-
-                if (isinstance(field, ForeignKey)):
-                    if field.unique:
-                        sql_unique_fields.append(name)
-
-                    field_name, field_id = field.model.get_id()
-                    fields.insert(0, (name, field_id.type_sql(dialect)))
-                    fields.append(('', field.sql(dialect, name)))
-                elif (isinstance(field, ManyToOne)):
+                from duck_orm.sql.relationship import OneToOne
+                if (isinstance(field, OneToOne)):
                     field_id = field.model.get_id()[1]
-                    fields.append((name, field_id.type_sql(dialect)))
-                elif (isinstance(field, OneToMany)):
-                    continue
-                elif (isinstance(field, OneToOne)):
-                    field_name, field_id = field.model.get_id()
                     table_name = field.model.get_name()
                     fields.append((name, field_id.column_sql(dialect)))
-                    fields.append(('', field.sql(dialect, name, table_name)))
+                    fields.append(
+                        ('', field.create_sql(dialect, name, table_name)))
                 else:
                     fields.insert(0, (name, field.column_sql(dialect)))
 
@@ -125,7 +136,7 @@ class Model:
                 if (field.primary_key):
                     return name, field
 
-        raise Exception('Model has no primary key!')
+        raise IdInvalidException('Model has no primary key!')
 
     @ classmethod
     def __get_select_sql(
@@ -157,7 +168,11 @@ class Model:
         fields_foreign_key: Dict[str, OneToMany] = {}
         for name, field in inspect.getmembers(cls):
             if isinstance(field, fields_type.Column):
-                fields_all.append(name)
+                if isinstance(field, fields_type.Timestamp):
+                    fields_all.append((name, field))
+                else:
+                    fields_all.append(name)
+
                 if isinstance(field, (ManyToOne, OneToOne, ForeignKey)):
                     field_name = field.model.get_id()[0]
                     condition_with_id = Condition(field_name, '=', data[name])
@@ -182,7 +197,7 @@ class Model:
         result: List[cls] = []
         dialect = get_dialect(str(cls.__db__.url.dialect))
         for row in data:
-            row = dict(row.items())
+            row = dict(row._mapping.items())
             fields_all, fields_foreign_key = await cls.__parser_fields(row)
             entity = dialect.parser(row, fields_all, fields_foreign_key)
             result.append(cls(**entity))
@@ -202,10 +217,26 @@ class Model:
         dialect = get_dialect(str(cls.__db__.url.dialect))
         result: cls = None
         if (data is not None):
-            data = dict(data.items())
+            data = dict(data._mapping.items())
             fields_all, fields_foreign_key = await cls.__parser_fields(data)
             entity = dialect.parser(data, fields_all, fields_foreign_key)
             result: cls = cls(**entity)
+        return result
+
+    @classmethod
+    async def find_by_id(
+        cls: Type[T],
+        id: Any,
+        fields_includes: List[str] = [],
+        fields_excludes: List[str] = []
+    ) -> Type[T]:
+        name = cls.get_id()[0]
+        condition = Condition(name, "=", id)
+        result = await cls.find_one(
+            fields_includes=fields_includes,
+            fields_excludes=fields_excludes,
+            conditions=[condition]
+        )
         return result
 
     @ classmethod
@@ -215,7 +246,7 @@ class Model:
         data = await cls.__db__.fetch_all(sql)
         result: List = []
         for row in data:
-            row = dict(row.items())
+            row = dict(row._mapping.items())
             entity = query_executor.parser(row)
             result.append(entity)
 
@@ -259,7 +290,7 @@ class Model:
         data = await model.__get_last_insert_id()
         name, field = cls.get_id()
         if (data is not None):
-            data = dict(data.items())
+            data = dict(data._mapping.items())
             from duck_orm.sql.relationship import OneToOne
             if isinstance(field, OneToOne):
                 field_model = model._instance[name]
@@ -305,13 +336,15 @@ class Model:
         return await self.find_one(conditions=[condition_with_id])
 
     @ classmethod
-    def __drop_table(cls, name_table: str, dialect: str):
+    def __drop_table(
+            cls, dialect: str, name_table: str, cascade: bool = False):
         query_executor = get_dialect(dialect)
-        return query_executor.drop_table(name_table)
+        return query_executor.drop_table(name_table, cascade)
 
     @ classmethod
-    async def drop_table(cls):
-        sql = cls.__drop_table(cls.get_name(), str(cls.__db__.url.dialect))
+    async def drop_table(cls, cascade: bool = False):
+        sql = cls.__drop_table(
+            str(cls.__db__.url.dialect), cls.get_name(), cascade)
         await cls.__db__.execute(sql)
 
     @ classmethod