+++ tests/test_duck_sqlite.py
@@ -4,6 +4,7 @@ import asyncio
 import pytest
 
 from duck_orm.model import Model
+from duck_orm.model_manager import ModelManager
 from duck_orm.sql import fields as Field
 from duck_orm.sql.condition import Condition
 from duck_orm.exceptions import UpdateException
@@ -11,9 +12,12 @@ from duck_orm.sql.relationship import ForeignKey
 
 db = Database('sqlite:///example.db')
 
+model_manager = ModelManager()
+
 
 class MyTest(Model):
     __db__ = db
+    model_manager = model_manager
 
     id: int = Field.Integer(primary_key=True, auto_increment=True)
     msg: str = Field.String(not_null=True)
@@ -22,25 +26,33 @@ class MyTest(Model):
 class Person(Model):
     __tablename__ = 'persons'
     __db__ = db
+    model_manager = model_manager
 
     id: int = Field.Integer(primary_key=True, auto_increment=True)
     first_name: str = Field.String(unique=True)
     last_name: str = Field.String(not_null=True)
     age: int = Field.Integer()
     salary: int = Field.BigInteger()
+    alive: bool = Field.Boolean()
 
 
 class Son(Model):
     __tablename__ = 'sons'
     __db__ = db
+    model_manager = model_manager
 
     id: int = Field.Integer(primary_key=True, auto_increment=True)
+    description: str = Field.Varchar(
+        length=15, default_value="Has no description")
     first_name: str = Field.String(unique=True)
     last_name: str = Field.String(not_null=True)
     age: int = Field.Integer()
-    person_id: int = ForeignKey(
-        model=Person, name_in_table_fk="id",
-        on_delete="NO ACTION", on_update="CASCADE")
+
+    @classmethod
+    def relationships(cls):
+        cls.person_id: int = ForeignKey(
+            model=Person, name_in_table_fk="id",
+            on_delete="CASCADE", on_update="CASCADE")
 
 
 def async_decorator(func):
@@ -72,19 +84,18 @@ def test_create_sql():
         "last_name TEXT NOT NULL, " + \
         "id INTEGER PRIMARY KEY AUTOINCREMENT, " + \
         "first_name TEXT UNIQUE, " + \
+        "alive INTEGER, " + \
         "age INTEGER);"
 
 
 def test_create_sql_son():
     sql = Son._Model__get_create_sql()
     assert sql == "CREATE TABLE IF NOT EXISTS sons (" + \
-        "person_id INTEGER, " + \
         "last_name TEXT NOT NULL, " + \
         "id INTEGER PRIMARY KEY AUTOINCREMENT, " + \
         "first_name TEXT UNIQUE, " + \
-        "age INTEGER,  " + \
-        "FOREIGN KEY (person_id) REFERENCES persons (id) " + \
-        "ON DELETE NO ACTION ON UPDATE CASCADE);"
+        "description VARCHAR(15) DEFAULT 'Has no description', " + \
+        "age INTEGER);"
 
 
 def get_table(table, tables):
@@ -97,9 +108,7 @@ def get_table(table, tables):
 @async_decorator
 async def test_create_table():
     await db.connect()
-    await Person.create()
-    await Son.create()
-    await MyTest.create()
+    await model_manager.create_all_tables()
     tables = await Person.find_all_tables()
     assert get_table('persons', tables)
     assert get_table('mytest', tables)
@@ -155,7 +164,7 @@ async def test_sql_select_where_persons():
     assert fields.__contains__('first_name')
     assert fields.__contains__('last_name')
     assert fields.__contains__('salary')
-    msg = "SELECT {fields} FROM persons WHERE first_name = 'Rich';".format(
+    msg = "SELECT {fields} FROM persons WHERE first_name = 'Rich'".format(
         fields=fields)
     assert sql[0] == msg
 
@@ -180,6 +189,21 @@ async def test_select_all_limit():
     assert len(persons) == 2
     assert persons[0].first_name == 'Rich'
     assert persons[1].first_name == 'Lucas'
+
+
+@async_decorator
+async def test_find_by_id_success():
+    person = await Person.find_by_id(1)
+    assert person.first_name == 'Rich'
+    assert person.last_name == 'Rich Ramalho'
+    assert person.age == 21
+    assert person.salary == 10000000
+
+
+@async_decorator
+async def test_find_by_id_invalid():
+    person = await Person.find_by_id(4)
+    assert person is None
 
 
 @async_decorator
@@ -250,6 +274,4 @@ async def test_update_sql_without_id():
 
 @async_decorator
 async def test_drop_table():
-    await Son.drop_table()
-    await Person.drop_table()
-    await MyTest.drop_table()
+    await model_manager.drop_all_tables()